Create the redux store
Import the fakeData.
Generate an array of movie objects. Make sure that all csv properties are converted to arrays.
Using a predefined array of filterTypes, generate a new array of filter objects, adding a property for being enabled.
Set the initial state of the store.
Create the store using the redux dev tools.